fix: 입력 받은 시도 횟수 검증
- 입력 값이 정수가 아닌 경우 IllegalArgumentException 발생
- 음수 입력 시 IllegalArgumentException 발생
- 입력 값이 null일 경우 IllegalArgumentException 발생

+++ src/main/kotlin/racingcar/Application.kt
@@ -6,5 +6,5 @@ fun main() {
 
     println("실행 결과")
 
-    winner(names, tryNumber.toInt())
+    winner(names, tryNumber)
 }
